Reorganize convert to links modal
